+++ lib/handler.js
@@ -1,7 +1,8 @@
-const AWS = require('aws-sdk')
 const utils = require('./utils')
 const uploadS3 = require('./upload-s3')
 const pgdump = require('./pgdump')
+const decorateWithIamToken = require('./iam')
+
 const encryption = require('./encryption')
 
 const DEFAULT_CONFIG = require('./config')
@@ -33,17 +34,6 @@ async function backup(config) {
     return uploadS3(stream, config, key)
 }
 
-function decorateWithIamToken(baseConfig) {
-    const rdsSigner = new AWS.RDS.Signer()
-    const token = rdsSigner.getAuthToken({
-        hostname: baseConfig.PGHOST,
-        port: baseConfig.PGPORT != null ? baseConfig.PGPORT : 5432,
-        region: baseConfig.S3_REGION,
-        username: baseConfig.PGUSER
-    })
-    return { ...baseConfig, PGPASSWORD: token }
-}
-
 async function handler(event) {
     const baseConfig = { ...DEFAULT_CONFIG, ...event }
     const config = event.USE_IAM_AUTH === true ? decorateWithIamToken(baseConfig) : baseConfig
@@ -59,4 +49,4 @@ async function handler(event) {
     }
 }
 
-module.exports = handler
+module.exports = handler, decorateWithIamToken
